html/browsers/browsing-the-web/navigating-across-documents/source/navigate-child-function.html

<!doctype html>
<meta charset=utf-8>
<title>Set location from a function called from a parent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe src="resources/location-set.html"></iframe>
<script>
async_test(t => {
  self.onload = t.step_func(() => {
    const fr = document.querySelector("iframe");
    const url = fr.contentDocument.URL;
    fr.contentWindow.go();
    fr.onload = t.step_func_done(() => {
      assert_equals(fr.contentDocument.URL, new URL("resources/dummy.html", self.location).href);
      assert_equals(fr.contentDocument.referrer, url);
    });
  });
});
</script>
